IftU: Add legacy translations
Since I couldn't really track down the original .po catalogues, I used
msgunfmt and msgmerge to decompile the .mo versions and merge them with
the up-to-date (well, 2.0.1) pot. It turns out the attribution
information is preserved in the .mo as part of the translation for msgid
"" (the empty string), so although fuzzies and untranslated strings are
dropped, it's not an issue for this particular case.

Of course, most strings have changed since then, so they're probably not
going to be very useful going forward, but they can serve as a starting
point for future translators I reckon.